LayoutTests/media/modern-media-controls/media-controller/media-controller-ios-only-enable-tap-gesture-recognizer-with-fades-when-idle-expected.txt

Testing the MediaController controls bar tap gesture recognizer is only enabled when fadesWhileIdle is true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default state
PASS controlsBar.fadesWhileIdle is false
PASS controlsBar._tapGestureRecognizer.enabled is false

controlsBar.fadesWhileIdle = true
PASS controlsBar.fadesWhileIdle is true
PASS controlsBar._tapGestureRecognizer.enabled is true

controlsBar.fadesWhileIdle = false
PASS controlsBar.fadesWhileIdle is false
PASS controlsBar._tapGestureRecognizer.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
